+++ packages/x/components/sender/context.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+import type { SenderProps } from './interface';
+
+export const SenderContext = React.createContext<SenderProps>({});
