+++ packages/x/components/sender/index.tsx
@@ -1,4 +1,4 @@
-import { Flex, Input } from 'antd';
+import { Flex } from 'antd';
 import classnames from 'classnames';
 import { useMergedState } from 'rc-util';
 import React from 'react';
@@ -6,149 +6,39 @@ import useProxyImperativeHandle from '../_util/hooks/use-proxy-imperative-handle
 import useXComponentConfig from '../_util/hooks/use-x-component-config';
 import { useXProviderContext } from '../x-provider';
 import SenderHeader, { SendHeaderContext } from './SenderHeader';
+import SlotTextArea, { type SlotTextAreaRef } from './SlotTextArea';
 import TextArea, { type TextAreaRef } from './TextArea';
 import { ActionButtonContext } from './components/ActionButton';
 import ClearButton from './components/ClearButton';
 import LoadingButton from './components/LoadingButton';
 import SendButton from './components/SendButton';
 import SpeechButton from './components/SpeechButton';
+import { SenderContext } from './context';
+import type {
+  ActionsComponents,
+  ActionsRender,
+  FooterRender,
+  SenderComponents,
+  SenderProps,
+  SenderRef,
+  SlotConfigType,
+  SubmitType,
+} from './interface';
 import useStyle from './style';
-import useSpeech, { type AllowSpeech } from './useSpeech';
+import useSpeech from './useSpeech';
 
-import type { ButtonProps, GetProps, InputProps } from 'antd';
-import SlotTextArea, { SlotTextAreaRef } from './SlotTextArea';
-
-type TextareaProps = GetProps<typeof Input.TextArea>;
-
-export type SubmitType = 'enter' | 'shiftEnter' | false;
-
-export interface SenderComponents {
-  input?: React.ComponentType<TextareaProps>;
-}
-
-type ActionsComponents = {
-  SendButton: React.ComponentType<ButtonProps>;
-  ClearButton: React.ComponentType<ButtonProps>;
-  LoadingButton: React.ComponentType<ButtonProps>;
-  SpeechButton: React.ComponentType<ButtonProps>;
+export type {
+  ActionsComponents,
+  ActionsRender,
+  FooterRender,
+  SenderComponents,
+  SenderProps,
+  SenderRef,
+  SlotConfigType,
+  SubmitType,
 };
 
-export type ActionsRender = (
-  ori: React.ReactNode,
-  info: {
-    components: ActionsComponents;
-  },
-) => React.ReactNode;
-
-export type FooterRender = (info: { components: ActionsComponents }) => React.ReactNode;
-
-interface SlotConfigBaseType {
-  type: 'text' | 'input' | 'select' | 'tag' | 'custom';
-  formatResult?: (value: any) => string;
-}
-
-interface SlotConfigTextType extends SlotConfigBaseType {
-  type: 'text';
-  text?: string;
-  key?: string;
-}
-
-interface SlotConfigInputType extends SlotConfigBaseType {
-  type: 'input';
-  key: string;
-  props?: {
-    defaultValue?: InputProps['defaultValue'];
-    placeholder?: string | undefined;
-  };
-}
-
-interface SlotConfigSelectType extends SlotConfigBaseType {
-  type: 'select';
-  key: string;
-  props?: {
-    defaultValue?: string;
-    options: string[];
-    placeholder?: string | undefined;
-  };
-}
-
-interface SlotConfigTagType extends SlotConfigBaseType {
-  type: 'tag';
-  key: string;
-  props?: {
-    label: React.ReactNode;
-    value?: string;
-  };
-}
-
-interface SlotConfigCustomType extends SlotConfigBaseType {
-  type: 'custom';
-  key: string;
-  props?: {
-    defaultValue?: any;
-    [key: string]: any;
-  };
-  customRender?: (
-    value: any,
-    onChange: (value: any) => void,
-    item: SlotConfigType,
-  ) => React.ReactNode;
-}
-
-export type SlotConfigType =
-  | SlotConfigTextType
-  | SlotConfigInputType
-  | SlotConfigSelectType
-  | SlotConfigTagType
-  | SlotConfigCustomType;
-
-export interface SenderProps
-  extends Pick<TextareaProps, 'placeholder' | 'onKeyUp' | 'onFocus' | 'onBlur'> {
-  prefixCls?: string;
-  defaultValue?: string;
-  value?: string;
-  loading?: boolean;
-  readOnly?: boolean;
-  submitType?: SubmitType;
-  disabled?: boolean;
-  slotConfig?: SlotConfigType[];
-  onSubmit?: (message: string, slotConfig?: SlotConfigType[]) => void;
-  onChange?: (
-    value: string,
-    event?: React.FormEvent<HTMLTextAreaElement> | React.ChangeEvent<HTMLTextAreaElement>,
-    slotConfig?: SlotConfigType[],
-  ) => void;
-  onCancel?: VoidFunction;
-  onKeyDown?: React.KeyboardEventHandler<any>;
-  onPaste?: React.ClipboardEventHandler<HTMLElement>;
-  onPasteFile?: (firstFile: File, files: FileList) => void;
-  components?: SenderComponents;
-  styles?: {
-    prefix?: React.CSSProperties;
-    input?: React.CSSProperties;
-    actions?: React.CSSProperties;
-    footer?: React.CSSProperties;
-  };
-  rootClassName?: string;
-  classNames?: {
-    prefix?: string;
-    input?: string;
-    actions?: string;
-    footer?: string;
-  };
-  style?: React.CSSProperties;
-  className?: string;
-  actions?: React.ReactNode | ActionsRender;
-  allowSpeech?: AllowSpeech;
-  prefix?: React.ReactNode;
-  footer?: React.ReactNode | FooterRender;
-  header?: React.ReactNode;
-  autoSize?: boolean | { minRows?: number; maxRows?: number };
-}
-
-export const SenderContext = React.createContext<SenderProps>({});
-
-export type SenderRef = TextAreaRef | SlotTextAreaRef;
+export { SenderContext };
 
 /** Used for actions render needed components */
 const sharedRenderComponents = {
